Remove generic param from expand Replace panic with Result return in expand Move tests to own directory for better std handling Add InvalidLength error type with Error impl behind std feature gate Remove allocation in example
