+++ hkdf/examples/main.rs
@@ -11,13 +11,13 @@ fn main() {
     let info = hex::decode("f0f1f2f3f4f5f6f7f8f9").unwrap();
 
     let hk = Hkdf::<Sha256>::extract(Some(&salt[..]), &ikm);
-    let mut okm = vec![0u8; 42];
-    hk.expand(&info, &mut okm);
+    let mut okm = [0u8; 42];
+    hk.expand(&info, &mut okm).unwrap(); //Will never fail as 42 is a valid length for Sha256 to output
 
     println!("Vector 1 PRK is {}", hex::encode(hk.prk));
-    println!("Vector 1 OKM is {}", hex::encode(&okm));
+    println!("Vector 1 OKM is {}", hex::encode(&okm[..]));
     println!("Matched with https://tools.ietf.org/html/rfc5869#appendix-A.1");
 
     let expected = "3cb25f25faacd57a90434f64d0362f2a2d2d0a90cf1a5a4c5db02d56ecc4c5bf34007208d5b887185865";
-    assert_eq!(hex::encode(&okm), expected);
+    assert_eq!(hex::encode(&okm[..]), expected);
 }
